Updated ui base interface

+++ src/main/java/com/apokk/ui/UIElement.java
@@ -4,7 +4,7 @@ public interface UIElement {
     void draw();
     void update();
     void init();
-    void target(double value);
+    void target(float value);
     void setFailureMode(int mode);
     String getID();
 }
